page objects divided to more obvious, gherkin fixed
